pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "wikiscraper"
version = "0.1.0"
description = "Modular Wikipedia scraper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "pandas>=3.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "wordfreq>=3.1.1",
]
[project.scripts]
wikiscraper = "wikiscraper.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
python_files = ["unit_tests.py"]


[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = []
